consolidate structures models
