--- nanome/_internal/structure/base.py
@@ -1,7 +1,0 @@
-__metaclass__ = type
-
-
-class _Base(object):
-    def __init__(self):
-        self._index = -1
-        self._parent = None
