--- nanome/_internal/structure/workspace.py
@@ -1,55 +1,0 @@
-from . import _Base
-from nanome._internal.util.decorators import deprecated
-
-
-class _Workspace(_Base):
-
-    @classmethod
-    def _create(cls):
-        return cls()
-
-    def __init__(self):
-        from nanome.util import Vector3, Quaternion
-        self._position = Vector3()
-        self._rotation = Quaternion()
-        self._scale = Vector3(0.02, 0.02, 0.02)
-        self._complexes = []
-
-    def _add_complex(self, complex):
-        self._complexes.append(complex)
-        complex._parent = self
-
-    def _remove_complex(self, complex):
-        if complex in self._complexes:
-            self._complexes.remove(complex)
-            complex._parent = None
-
-    @deprecated()
-    def get_atom_iterator(self):
-        iterator = _Workspace.AtomIterator(self)
-        return iter(iterator)
-
-    class AtomIterator(object):
-        def __init__(self, workspace):
-            self._workspace = workspace
-
-        def __iter__(self):
-            self._complexes = iter(self._workspace.complexes)
-            self._update_iter()
-            return self
-
-        def __next__(self):
-            while True:
-                try:
-                    return next(self._moleculeAtom)
-                except StopIteration:
-                    self._update_iter()
-
-        def _update_iter(self):
-            while True:
-                complex = next(self._complexes)
-                try:
-                    self._moleculeAtom = complex.get_atom_iterator()
-                    break
-                except StopIteration:
-                    pass
